# pkg/gen/models/party.go
// Code generated by go-swagger; DO NOT EDIT.

package models

// This file was generated by the swagger tool.
// Editing this file might prove futile when you re-run the swagger generate command

import (
	strfmt "github.com/go-openapi/strfmt"

	"github.com/go-openapi/swag"
)

// Party party
// swagger:model Party
type Party struct {

	// account name
	AccountName string `json:"account_name,omitempty"`

	// account number
	AccountNumber string `json:"account_number,omitempty"`
}

// Validate validates this party
func (m *Party) Validate(formats strfmt.Registry) error {
	return nil
}

// MarshalBinary interface implementation
func (m *Party) MarshalBinary() ([]byte, error) {
	if m == nil {
		return nil, nil
	}
	return swag.WriteJSON(m)
}

// UnmarshalBinary interface implementation
func (m *Party) UnmarshalBinary(b []byte) error {
	var res Party
	if err := swag.ReadJSON(b, &res); err != nil {
		return err
	}
	*m = res
	return nil
}